requirements.txt
pymupdf
streamlit
easyocr
numpy
python-dotenv
groq
chromadb
sentence-transformers
python-multipart
fastapi
uvicorn[standard]